qb-jobs Support Added
Added support for the qb-jobs manager menu.

move the getJobs function to above setCityHallPageState function as it should be called every time before the main menu opens or the variable will be out of date.

app.js

set it up to open the name.label and to run the setjobs function when the open object is accessed.

removed height from the style.css to allow for asthetic reasons with the menu and makes it scroll better.

languages

Added a few language variables to support the new qb-jobs application system.