Lots of changes something something something QBJobs(#88)
* changes for qb-jobs

Ready for qb-jobs

* Update main.lua

made linter happy

* Update main.lua

removed the table check for availableJobs in AddCityJob function.

* Fixed Issues with AddCityJob export.

Worked with Kody on solving the issue with the dang blasted export not working until server restart.

We got it functional and it's included in this PR now.

* stuff requested and required

fixed linting
appeased stan! :)

* Update main.lua

made charj happy

* Update main.lua

Added UpdateObject to work with qb-jobs export

* Update style.css

Created Compatibility with new qb-jobs management menu.

* qb-jobs Support Added

Added support for the qb-jobs manager menu.

move the getJobs function to above setCityHallPageState function as it should be called every time before the main menu opens or the variable will be out of date.

app.js

set it up to open the name.label and to run the setjobs function when the open object is accessed.

removed height from the style.css to allow for asthetic reasons with the menu and makes it scroll better.

languages

Added a few language variables to support the new qb-jobs application system.

* Created backwards compatibility

I added the if conditional to run the export if qb-jbs is installed or the default method when not installed. Works on vanilla from my test.

* removed unused function

removed the sendemail function

* qb-jobs temp

qb-jobs temp

* Revert "qb-jobs temp"

This reverts commit 875e6508cb9b4c8fdaa9ae321e659ba225b53966.

* Update app.js

Prevents duplicate available jobs from appearing upon restart of resource that exports to cityhall.

* Bump

---------

Co-authored-by: zwkf <dev@charlie.fyi>